Fix sanitizer overflow crash in mp4 extractor.
mp4 extractor experienced a sanitizer overflow error due to an
overflow when calculating the actual seek position based on the
track playback offset.
The problematic video has an abnormally high offset value.

Referencing media3 to implement rescaleTime: ensure that precision
is not lost during the calculation and to handle overflow cases.

Bug: 397619415
Test: atest MediaExtractorTest

Google: 3500491
Change-Id: Ieb7ad96dde6d345a306083bb60aa754aaff4f97e
Signed-off-by: Fang Zhaohui <fangzhaohui@xiaomi.corp-partner.google.com>
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>