AudioPlaybackCapture: solve patchTrack unable to update speed issue
PatchRecord is in the output thread of audiohal. Every time the track
is consumed by the mixer and the ReleaseBuffer is executed, the data
obtainBuffer is written into the PatchBuffer. When a third-party application
uses AudioTrack::setPlaybackRate to timestretch the audio data, PatchRecord
and PatchTrack need to update framecount and speed at the same time. And the
remotesubmix thread where PatchTrack is located needs to perform timestretch
processing synchronously to solve the problem of Track::interceptBuffer
dropping frames.

Signed-off-by: v-chongyin@xiaomi.com
Google: 3586731
Change-Id: I6b831b5a86ead8d940b9e0f474268c750b01592c
Signed-off-by: Edwiin Kusuma Jaya <kutemeikito0905@gmail.com>